èp34 继集成mybatis generator

+++ src/main/java/life/majiang/community/Controller/PublishController.java
@@ -48,7 +48,7 @@ public class PublishController {
             @RequestParam(value = "title",required = false) String title,
             @RequestParam(value="description",required = false) String description,
             @RequestParam(value = "tag",required = false) String tag,
-            @RequestParam(value = "id",required = false) Long id,
+            @RequestParam(value = "id",required = false) Integer id,
             HttpServletRequest httpServletRequest,
             Model model
     ) {
