+++ src/main/java/life/majiang/community/dto/QuestionDto.java
@@ -9,11 +9,11 @@ import lombok.Data;
 */
 @Data
 public class QuestionDto {
-    private  long id;
+    private  Integer id;
     private String title;
     private Long gmtCreate;
     private Long gmtModified;
-    private Long creator;
+    private Integer creator;
     private Integer commentCount;
     private Integer viewCount;
     private Integer likeCount;
